Add unit tests for migrating declaration.
Add unit tests for migrating declaration from the old-style
XML language to the new-style YAML language.

Refs #113677

+++ wres-config/src/wres/config/yaml/components/SpatialMask.java
@@ -15,4 +15,4 @@ import wres.config.yaml.deserializers.SpatialMaskDeserializer;
 @JsonDeserialize( using = SpatialMaskDeserializer.class )
 public record SpatialMask( @JsonProperty( "name" ) String name,
                            @JsonProperty( "wkt" ) String wkt,
-                           @JsonProperty( "srid" ) Integer srid ) {}
+                           @JsonProperty( "srid" ) Long srid ) {}
